backend/pyproject.toml
[project]
name = "eo-cd-slo-backend"
version = "1.0.0"
description = "Sentinel-2 Earth Observation Change Detection API Backend"
authors = [{name = "Your Name", email = "your.email@example.com"}]
requires-python = ">=3.10,<3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "asyncpg==0.29.0",
    "sqlalchemy[asyncio]==2.0.23",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "Pillow==10.1.0",
    "rasterio==1.3.9",
    "numpy==1.24.3",
    "aiofiles==23.2.0",
    "logfire[fastapi]==3.21.1",
]

[tool.uv]
dev-dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
